расчет скидок, начисление бонусов

+++ data/schema.sql
@@ -6,7 +6,7 @@ USE coursework;
 CREATE TABLE goods (
   id int(11) NOT NULL PRIMARY KEY AUTO_INCREMENT,           -- id товара
   title varchar(60) NOT NULL,                               -- заголовок товара
-  price int(11) NOT NULL,                                   -- цена товара
+  price double NOT NULL,                                    -- цена товара
   description text NOT NULL,                                -- описание товара
   FULLTEXT (title)
 );
